- Fixed signing out: the problem was in StartView - there should be conditions so that normal HomeView and ProfileView could not load and deliver nil error. Moreover there should be dismissSheet method in ProfileView in DeleteAccountSheetView

+++ Date-Up/utils/FirebaseStorageManager.swift
@@ -50,15 +50,19 @@ class FirebaseStorageManager: ObservableObject {
     }
     
     func deleteAllImagesFromStorage(userID: String, userPhotosURLs: [String], completion: @escaping (() -> ())) {
-        for photoURL in userPhotosURLs {
-            let userImagesStorageRef = storageRef.child("images/\(userID)/\(photoURL)")
-            userImagesStorageRef.delete() { (error) in
-                if let error = error {
-                    print("Error deleting file: ", error.localizedDescription)
+        if userPhotosURLs.count != 0 {
+            for photoURL in userPhotosURLs {
+                let userImagesStorageRef = storageRef.child("images/\(userID)/\(photoURL)")
+                userImagesStorageRef.delete() { (error) in
+                    if let error = error {
+                        print("Error deleting file: ", error.localizedDescription)
+                    } else {
+                        completion()
+                    }
                 }
-                
-                completion()
             }
+        } else {
+            completion()
         }
     }
